requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
boto3==1.34.108
botocore==1.34.108
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.13
django-celery==3.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-import-export==4.0.3
django-node==4.0.0
django-storages==1.14.3
django-webpack==3.1.1
gunicorn==22.0.0
idna==3.7
jmespath==1.0.1
kombu==5.3.7
optional-django==0.1.0
packaging==24.0
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
requests==2.31.0
s3transfer==0.10.1
sentry-sdk==2.2.0
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
typing-extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
